:sparkles: Feat: refresh 로직 작성 [YW-149]
